✨ RENDERER: Local Distributed Rendering
💡 What: Implemented RenderOrchestrator to split render jobs into concurrent chunks and concatenate them. Refactored Renderer to its own file.
🎯 Why: To unlock multi-core rendering performance on a single machine.
📊 Impact: Faster rendering times for complex compositions.
🔬 Verification: Verified using packages/renderer/tests/verify-distributed.ts with concurrency: 2.
Reference: .sys/plans/2026-08-02-RENDERER-Local-Distributed-Rendering.md
