+++ .sys/llmdocs/context-renderer.md
@@ -13,6 +13,8 @@ The Renderer operates on a "Dual-Path" architecture to support different use cas
 
 Both strategies pipe frame data directly to an FFmpeg process via stdin ("Zero Disk I/O"), ensuring high performance and low latency. Audio tracks from Blob URLs are extracted to memory and also piped to FFmpeg via additional pipes, avoiding temporary files.
 
+The **RenderOrchestrator** enables Local Distributed Rendering by splitting a job into concurrent chunks and merging the results, utilizing multiple cores for faster processing.
+
 ## B. File Tree
 ```
 packages/renderer/
@@ -30,7 +32,9 @@ packages/renderer/
 │   │   ├── FFmpegInspector.ts  # Environment diagnostics
 │   │   ├── dom-scanner.ts      # Asset discovery
 │   │   └── blob-extractor.ts   # Blob URL extraction
-│   ├── index.ts                # Main Renderer class
+│   ├── Orchestrator.ts         # Distributed rendering logic
+│   ├── Renderer.ts             # Main Renderer class
+│   ├── index.ts                # Public exports
 │   └── types.ts                # Configuration interfaces
 ├── scripts/                    # Self-contained verification scripts (integration tests)
 │   ├── verify-cancellation.ts  # Render cancellation test
@@ -40,6 +44,7 @@ packages/renderer/
 │   └── ...                     # Other script-based tests
 └── tests/
     ├── run-all.ts              # Test runner (executes comprehensive suite)
+    ├── verify-distributed.ts   # Distributed rendering test
     ├── verify-browser-config.ts # Browser launch config test
     ├── verify-waapi-sync.ts    # CSS animation sync test
     ├── verify-seek-driver-determinism.ts # SeekDriver determinism test
@@ -68,6 +73,7 @@ The `RendererOptions` interface controls the render pipeline:
 - `fps`: Target frame rate.
 - `durationInSeconds`: Total length of the video (fallback if `frameCount` is not set).
 - `frameCount`: Exact number of frames to render (overrides `durationInSeconds`).
+- `startFrame`: Frame to start rendering from (for range/distributed rendering).
 - `mode`: `'dom'` or `'canvas'`.
 - `canvasSelector`: CSS selector to target the canvas element in `'canvas'` mode (default `'canvas'`).
 - `browserConfig`: Object to customize Playwright browser launch (`headless`, `args`, `executablePath`).
@@ -79,6 +85,9 @@ The `RendererOptions` interface controls the render pipeline:
 - `intermediateImageQuality`: JPEG quality (0-100) if format is jpeg.
 - `stabilityTimeout`: Timeout for frame stability (default 30000ms).
 - `inputProps`: Object injected into the page as `window.__HELIOS_PROPS__`.
+
+The `DistributedRenderOptions` interface extends `RendererOptions` with:
+- `concurrency`: Number of parallel workers (defaults to CPU count - 1).
 
 ## D. FFmpeg Interface
 The renderer spawns an FFmpeg process with the following key flags:
